+++ starbase-nexus-api/Models/Constructions/Ship/ViewShip.cs
@@ -67,7 +67,7 @@ namespace starbase_nexus_api.Models.Constructions.Ship
         /// <summary>
         /// in minutes
         /// </summary>
-        public int? FlightTime { get; set; }
+        public float? FlightTime { get; set; }
 
         public float? TotalMassWithoutCargo { get; set; }
